backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
psycopg2-binary
python-dotenv
python-multipart
groq
google-genai
passlib[bcrypt]
bcrypt==4.0.1
PyJWT
email-validator